Solve clearance command-query exercise. Inherited from Array class in clearance.rb to
make my Clearance class act like an array by default. Set up Item to pull values from the
key value pair arguments using 'params', which is a behind-the-scenes hash object created
for you when you pass in key value pairs as the last arguments in an arguments list. Initially
had best_deal as an instance variable, but switched to method to return nil if empty, and
get the best deal if not. Cleaner that way.